+++ client/src/main/java/net/oneandone/stool/client/cli/InfoCommand.java
@@ -90,7 +90,7 @@ public abstract class InfoCommand extends StageCommand {
         Project project;
         List<Reference> result;
 
-        project = Project.lookup(world.getWorking());
+        project = Project.lookup(projectDirectory);
         result = new ArrayList<>();
         if (project != null) {
             for (App app : project.list(serverManager)) {